+ CaminhaoService - deleteCaminhao refactor

+++ src/main/java/com/pedromiranda/transportadoraapi/service/CaminhaoService.java
@@ -33,6 +33,9 @@ public class CaminhaoService {
     }
 
     public void deleteCaminhao(Long id) {
+        if (repository.findById(id).isEmpty())
+            throw new CaminhaoNotFoundException(id);
+
         repository.deleteById(id);
     }
 }
